Setup.sh better support for only updating

+++ setup.sh
@@ -45,32 +45,34 @@ sudo apt install cmake -y
 sudo apt install swig -y
 fi
 
-displayMsg "Now, we use ${BLUE}git to clone necessary repositories${NC}."
+displayMsg "Now, we use ${BLUE}git to clone (or update) necessary repositories${NC}."
 
 if $execStep; then
-git clone https://github.com/orat/ConformalGeometricAlgebra
-git clone https://github.com/orat/GeometricAlgebra
-git clone https://github.com/orat/Euclid3DViewAPI
-git clone https://github.com/orat/EuclidView3d
-git clone https://github.com/orat/SparseMatrix
-git clone https://github.com/orat/CGACasADi
-git clone https://github.com/MobMonRob/JCasADi
-git clone https://github.com/MobMonRob/JNativeLibLoader
-git clone https://github.com/orat/GACalcAPI
+git clone https://github.com/orat/ConformalGeometricAlgebra &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/orat/GeometricAlgebra &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/orat/Euclid3DViewAPI &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/orat/EuclidView3d &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/orat/SparseMatrix &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/orat/CGACasADi &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/MobMonRob/JCasADi &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/MobMonRob/JNativeLibLoader &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
+git clone https://github.com/orat/GACalcAPI &> /dev/null || ( cd ConformalGeometricAlgebra && git pull && cd .. )
 wget -O vecmath.jar https://jogamp.org/deployment/java3d/1.7.1-build-20200222/vecmath.jar
 fi
 
 displayMsg "Now, we ${BLUE}download and install GraalVM and Maven${NC}."
 
 if $execStep; then
-wget -O graalvm21.tar.gz https://github.com/graalvm/graalvm-ce-builds/releases/download/jdk-21.0.2/graalvm-community-jdk-21.0.2_linux-x64_bin.tar.gz
-mkdir GraalVM21
-tar -xzf graalvm21.tar.gz -C ./GraalVM21 --strip-components=1
-rm -f graalvm21.tar.gz
-wget -O maven.tar.gz https://dlcdn.apache.org/maven/maven-3/3.9.9/binaries/apache-maven-3.9.9-bin.tar.gz
-mkdir Maven3.9.9
-tar -xzf maven.tar.gz -C ./Maven3.9.9 --strip-components=1
-rm -f maven.tar.gz
+if mkdir GraalVM21 &> /dev/null; then
+	wget -O graalvm21.tar.gz https://github.com/graalvm/graalvm-ce-builds/releases/download/jdk-21.0.2/graalvm-community-jdk-21.0.2_linux-x64_bin.tar.gz
+	tar -xzf graalvm21.tar.gz -C ./GraalVM21 --strip-components=1
+	rm -f graalvm21.tar.gz
+fi
+if mkdir Maven3.9.9 &> /dev/null; then
+	wget -O maven.tar.gz https://dlcdn.apache.org/maven/maven-3/3.9.9/binaries/apache-maven-3.9.9-bin.tar.gz
+	tar -xzf maven.tar.gz -C ./Maven3.9.9 --strip-components=1
+	rm -f maven.tar.gz
+fi
 cd ./GraalVM21
 export JAVA_HOME=$(pwd)
 cd ../Maven3.9.9
@@ -83,7 +85,7 @@ fi
 displayMsg "Finally, we ${BLUE}build the repositories with Maven${NC}."
 
 if $execStep; then
-mvn -v || { printf "\n${RED}Maven can't be executed${NC}. Try running the install script again without skipping the '${BLUE}download and install GraalVM and Maven${NC}' step or install it manually.\n"; exit 1; }
+mvn -v ||  { printf "\n${RED}Maven can't be executed${NC}. Try running the install script again without skipping the '${BLUE}download and install GraalVM and Maven${NC}' step or install it manually.\n"; exit 1; }
 mvnInstallCmd="mvn --no-transfer-progress install"
 cd ./SparseMatrix && $mvnInstallCmd || installationEval
 cd ../GeometricAlgebra && $mvnInstallCmd || installationEval
